Autoload TinyMCE templates for post types
TinyMCE templates is always underestimated but it's a good alternative
to page builders.
Thistle loads automaticaly TinyMCE templates plugin and looks at files
matching the pattern `tinymce-*.php` into the `template-parts` folder.
To be more friendly, the behaviour of TinnyMCE templates is the same as
page templates. They use a comment header with some needed information:
"Title" (required), "Description" and Post type to load on demand.
If the templates list is empty, the template plugin is pulled out, you
don't need to remove the template button ;)