Manage the robots.txt file content
Thistle tries to provide a better way to use this WP feature
with the help of a multidimensional array instead of an awkward string.
This behaviour comes at the end of the filter `robots_txt`
and unfortunately does not take into account all information added to
the initial string. It's uncommon that plugins add something into
robots.txt even if they should. For example, Google XML Sitemaps doesn't
use this filter and pass to the action `do_robotstxt` :/