Change the default version of dependencies
By default, all dependencies without defined version use
the WP one instead.
It's not a bad idea but as it's a good practice to hide the WP version,
it becomes a "security issue". To be sure to don't find WP version
into dependency URLs, we modify default dependency version at the source
to be sure to impact a maximum of cases
(e.g.: `_print_styles` from `wp-login.php`).