Remove the pagination permalink base
The pagination should not be a part of the URL but lives only
as a parameter.
All URLs like `archive/page/2/` or `postname/2/` will live through
the same GET parameter: `page`.

The behaviour of multiple pages is also enhanced for all post types.
Thistle provides an URL to read the whole post, adds a link "View all"
into the pagination and displays the relational links
for the pages adjacent.