Manage conflict with the new "View Archive" link
WordPress adds its own link into the admin bar to go to the archive page
of any post types. But its behaviour does not work correctly with
Thistle because of the possibility to have a post type without
single view (`publicly_queryable => false`).
So the Thistle link is kept in favour of the WP one.