Disable almost all of the REST API
Since WP 4.7, the REST API is enabled by default and
the content endpoints are not protected. This behaviour can be
considered as a security issue and needs to be fixed.
Most of plugins disable the REST API in full BUT we should be attentive
to some details. It's important to know that some features of
the WP Core like "Post embeds" or plugins (Yoast SEO amongst others) use
the REST API.
If you turn off the REST API, you will break them without knowing it.

Thistle takes another way which is softer. It disabled all routes of
the REST API for non logged users except the `/oembed/1.0/embed/` one.
With this behaviour, you can embed your posts on other WordPress sites
and use all your plugins which use the REST API without any problem ;)

https://wptavern.com/wordpress-rest-api-vulnerability-exploits-continue