Manage the medium_large_size_* settings
Adds new fields into the media settings page to manage
the fourth image size: "Medium Large".

Introduced by WordPress 4.4, this format has by default a `768px` width
and is used for responsive images (through the `srcset` attribute).
I know that it's not a good idea to define `srcset` and `size`
attributes according to this site breakpoint
(cf.: https://cloudfour.com/thinks/responsive-images-101-part-9-image-breakpoints/#winging-it-aka-matching-your-layout-breakpoints)
but as WP handles this format (and uses space disc) without telling us,
we should be able to edit it to enjoy it
even for one of our breakpoints.

If you want to escape this format, you just have to set its width
and height to zero ;)